Digital makeup/requirements.txt
appdirs==1.4.4
certifi==2020.12.5
click==7.1.2
distlib==0.3.1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
h5py==2.10.0
Keras==2.4.3
mpmath==1.2.1
mtcnn==0.1.0
ndindex==1.5.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pipenv==2020.11.15
PyYAML==5.4.1
scipy==1.6.2
six==1.15.0
sympy==1.7.1
versioned-hdf5==1.2.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
